refactor(@vllnt/ui): use direct sibling imports (no-barrel-import)
Replace 52 sibling-component barrel imports (e.g. ../badge) with direct
module paths (../badge/badge) across component files. Improves tree-shaking
and matches the repo's 'direct imports internal, barrel external' convention.

Scoped to the imports react-doctor explicitly flagged. Did not force
no-barrel-import to zero: converting every sibling barrel re-introduces a
circular-import cycle the barrel indirection guards (broke a tutorial-complete
render test). 26 advisory no-barrel-import warnings remain by design.

Part of #279
