+++ packages/ui/src/components/chat-dock-section/chat-dock-section.tsx
@@ -4,7 +4,7 @@ import { ArrowUpRight, MessageSquareText } from "lucide-react";
 import type { ReactNode } from "react";
 
 import { cn } from "../../lib/utils";
-import { Button } from "../button";
+import { Button } from "../button/button";
 
 export type ChatDockMessage = {
   body: ReactNode;
