+++ packages/ui/src/components/code-playground/code-playground.tsx
@@ -7,7 +7,7 @@ import type { ReactNode } from "react";
 
 import type { HeadingTag } from "../../lib/types";
 import { cn } from "../../lib/utils";
-import { Button } from "../button";
+import { Button } from "../button/button";
 
 type CodeLineProps = {
   highlightLines: number[];
