+++ packages/ui/src/components/exercise/exercise.tsx
@@ -7,7 +7,7 @@ import type { ReactNode } from "react";
 
 import type { HeadingTag } from "../../lib/types";
 import { cn } from "../../lib/utils";
-import { Button } from "../button";
+import { Button } from "../button/button";
 
 const difficultyConfig = {
   easy: { className: "text-green-600 dark:text-green-400", label: "Easy" },
