+++ packages/ui/src/components/file-upload/file-upload.tsx
@@ -5,7 +5,7 @@ import * as React from "react";
 import { FileUp, UploadCloud, X } from "lucide-react";
 
 import { cn } from "../../lib/utils";
-import { Button } from "../button";
+import { Button } from "../button/button";
 
 export type FileUploadProps = Omit<
   React.ComponentPropsWithoutRef<"input">,
