+++ packages/ui/src/components/flashcard/flashcard.tsx
@@ -6,7 +6,7 @@ import { RefreshCcw } from "lucide-react";
 import type { ReactNode } from "react";
 
 import { cn } from "../../lib/utils";
-import { Badge } from "../badge";
+import { Badge } from "../badge/badge";
 import { Button } from "../button";
 import { Card, CardContent, CardFooter, CardHeader, CardTitle } from "../card";
