+++ packages/ui/src/components/form/form.tsx
@@ -19,7 +19,7 @@ import {
 } from "react-hook-form";
 
 import { cn } from "../../lib/utils";
-import { Label } from "../label";
+import { Label } from "../label/label";
 
 type FormInstance<TFieldValues extends FieldValues> = UseFormReturn<
   TFieldValues,
