+++ packages/ui/src/components/inline-input/inline-input.tsx
@@ -3,7 +3,7 @@
 import type { KeyboardEvent } from "react";
 
 import { cn } from "../../lib/utils";
-import { Input } from "../input";
+import { Input } from "../input/input";
 
 export type InlineInputProps = {
   className?: string;
