+++ packages/ui/src/components/model-selector/model-selector.tsx
@@ -5,7 +5,7 @@ import { useMemo, useRef, useState } from "react";
 import { ArrowUpDown, Filter } from "lucide-react";
 
 import { cn } from "../../lib/utils";
-import { Badge } from "../badge";
+import { Badge } from "../badge/badge";
 import { Button } from "../button";
 import {
   Command,
