+++ packages/ui/src/components/multi-select/multi-select.tsx
@@ -5,7 +5,7 @@ import * as React from "react";
 import { Check, ChevronDown } from "lucide-react";
 
 import { cn } from "../../lib/utils";
-import { Badge } from "../badge";
+import { Badge } from "../badge/badge";
 import { Button } from "../button";
 import {
   Command,
