+++ packages/ui/src/components/object-card/object-card.tsx
@@ -3,7 +3,7 @@ import { forwardRef } from "react";
 import type { ReactNode } from "react";
 
 import { cn } from "../../lib/utils";
-import { Badge } from "../badge";
+import { Badge } from "../badge/badge";
 import { Button } from "../button";
 
 export type ObjectCardMetric = {
