+++ packages/ui/src/components/search-dialog/search-dialog.tsx
@@ -12,7 +12,7 @@ import {
 import { Search } from "lucide-react";
 
 import { cn } from "../../lib/utils";
-import { Button } from "../button";
+import { Button } from "../button/button";
 import {
   CommandDialog,
   CommandEmpty,
