+++ packages/ui/src/components/sidebar-toggle/sidebar-toggle.tsx
@@ -3,7 +3,7 @@
 import { Menu, X } from "lucide-react";
 
 import { cn } from "../../lib/utils";
-import { Button } from "../button";
+import { Button } from "../button/button";
 
 export type SidebarToggleProps = {
   className?: string;
