+++ packages/ui/src/components/tutorial-card/tutorial-card.tsx
@@ -5,7 +5,7 @@ import { memo, useEffect, useState } from "react";
 import type { ReactNode } from "react";
 
 import { useMounted } from "../../lib/use-mounted";
-import { Badge } from "../badge";
+import { Badge } from "../badge/badge";
 import {
   Card,
   CardContent,
