+++ packages/ui/src/components/tutorial-filters/tutorial-filters.tsx
@@ -2,7 +2,7 @@
 
 import { memo } from "react";
 
-import { Badge } from "../badge";
+import { Badge } from "../badge/badge";
 
 export type TutorialFiltersLabels = {
   activeFilters: string;
