+++ packages/ui/src/components/world-clock-bar/world-clock-bar.tsx
@@ -4,7 +4,7 @@ import * as React from "react";
 
 import type { HeadingTag } from "../../lib/types";
 import { cn } from "../../lib/utils";
-import { Badge } from "../badge";
+import { Badge } from "../badge/badge";
 
 export type WorldClockBarZone = {
   city: string;
